chore(ts-sdk): expose `pureBcsSchemaFromTypeName` utility (#10161)
ports commits:

https://github.com/MystenLabs/ts-sdks/commit/3d8a0d9c82d69cf8ea13302ba33a8ae87c8891e3
https://github.com/MystenLabs/ts-sdks/commit/5dfe70881edd40f16dbfa47f20634634277f8f64

Related to: https://github.com/iotaledger/iota/issues/10144
