+++ .changeset/fifty-cats-allow.md
@@ -1,0 +1,5 @@
+---
+'@iota/iota-sdk': minor
+---
+
+expose `pureBcsSchemaFromTypeName` utility
